Initialize CVS branch netbsd-3-1
The per-file CVS branch roots do not exactly match a reconstructed source-branch commit.

+++ distrib/utils/sysinst/net.c
@@ -1,4 +1,4 @@
-/*	$NetBSD: net.c,v 1.106 2005/05/08 19:46:33 christos Exp $	*/
+/*	$NetBSD: net.c,v 1.105.2.2 2006/09/16 10:26:09 ghen Exp $	*/
 
 /*
  * Copyright 1997 Piermont Information Systems Inc.
@@ -298,7 +298,7 @@ get_ifconfig_info(void)
 }
 
 static int
-do_ifreq(struct ifmediareq *ifmr, int cmd)
+do_ifreq(struct ifmediareq *ifmr, unsigned long cmd)
 {
 	int sock;
 	int rval;
